add docker building and running

+++ alphazero.h
@@ -10,6 +10,7 @@
 #include <atomic>
 #include <cassert>
 #include <cmath>
+#include <condition_variable>
 #include <mutex>
 #include <random>
 
@@ -21,7 +22,7 @@ template < typename MoveT, typename StateT > struct Payload
     using node_type = Node< MoveT, StateT, Payload >;
 
     Payload() noexcept = default;
-    Payload( Payload const &payload ) noexcept
+    Payload( Payload const& payload ) noexcept
         : policies_are_evaluated( payload.policies_are_evaluated.load() ),
           visits( payload.visits.load() ),
           nn_policy( payload.nn_policy.load() ),
@@ -63,17 +64,17 @@ struct GamePlay
 
 // not thread-safe.
 template < typename MoveT, typename StateT, typename PayloadT >
-float shannon_entropy( Node< MoveT, StateT, PayloadT > const &node )
+float shannon_entropy( Node< MoveT, StateT, PayloadT > const& node )
 {
     size_t visit_sum = 0;
-    for ( auto const &child : node.get_children() )
+    for ( auto const& child : node.get_children() )
         visit_sum +=
             child.get_payload().visits.load( std::memory_order_relaxed );
     if ( !visit_sum )
         return 0.0f;
 
     float entropy = 0;
-    for ( auto const &child : node.get_children() )
+    for ( auto const& child : node.get_children() )
     {
         const size_t visits =
             child.get_payload().visits.load( std::memory_order_relaxed );
@@ -81,7 +82,7 @@ float shannon_entropy( Node< MoveT, StateT, PayloadT > const &node )
             continue;
         const float p =
             static_cast< float >( visits ) / static_cast< float >( visit_sum );
-        entropy -= p * std::logf( p );
+        entropy -= p * std::log( p );
     }
     return entropy;
 }
@@ -109,11 +110,11 @@ class Player : public ::Player< MoveT >
     static constexpr std::size_t game_size = G;
     static constexpr std::size_t policy_size = P;
 
-    Player( game_type game, params::Ucb const &ucb,
-            params::GamePlay const &game_play,
+    Player( game_type game, params::Ucb const& ucb,
+            params::GamePlay const& game_play,
             unsigned seed, // make the play deterministic with seed
-            allocator_type &allocator,
-            inference_service_type &inference_service )
+            allocator_type& allocator,
+            inference_service_type& inference_service )
         : allocator( allocator ), ucb_params( ucb ), game_play( game_play ),
           g( seed ), thread_pool( game_play.parallel_simulations ),
           inference_service( inference_service ),
@@ -126,7 +127,7 @@ class Player : public ::Player< MoveT >
         if ( !game_play.parallel_simulations )
             throw beat_it::Exception( "Parallel simulations cannot be zero." );
 
-        for ( auto &thread : thread_pool )
+        for ( auto& thread : thread_pool )
             thread =
                 std::jthread( &Player::worker, this, stop_source.get_token() );
     }
@@ -136,27 +137,27 @@ class Player : public ::Player< MoveT >
         stop_source.request_stop();
         // release all potentially waiting workers
         remaining_simulations.release( thread_pool.size() );
-        for ( auto &thread : thread_pool )
+        for ( auto& thread : thread_pool )
             thread.join();
     }
 
     // not thread safe.
     MoveT choose_move() override
     {
-        auto &node = run_simulations();
+        auto& node = run_simulations();
         allocator.reset();
         root = node.copy_tree( allocator );
         return root.get().get_move();
     }
 
-    node_type &get_root() const noexcept { return root; }
+    node_type& get_root() const noexcept { return root; }
 
     // require: move is valid, not thread safe.
-    void apply_opponent_move( MoveT const &move ) override
+    void apply_opponent_move( MoveT const& move ) override
     {
-        auto &children = root.get().get_children();
+        auto& children = root.get().get_children();
         auto new_root_itr =
-            std::ranges::find_if( children, [move]( auto const &node )
+            std::ranges::find_if( children, [move]( auto const& node )
                                   { return node.get_move() == move; } );
 
         allocator.reset();
@@ -166,7 +167,7 @@ class Player : public ::Player< MoveT >
             return;
         }
 
-        auto *pre_node = dynamic_cast< pre_node_type * >( &root.get() );
+        auto* pre_node = dynamic_cast< pre_node_type* >( &root.get() );
         if ( pre_node )
         {
             root = *(
@@ -180,24 +181,24 @@ class Player : public ::Player< MoveT >
 
   protected:
     // promise: return index of move in policy_vector
-    virtual size_t move_to_policy_index( MoveT const & ) const = 0;
+    virtual size_t move_to_policy_index( MoveT const& ) const = 0;
     virtual std::array< float, G >
-    serialize_state( game_type const & ) const = 0;
+    serialize_state( game_type const& ) const = 0;
 
   private:
     friend class training::SelfPlay< MoveT, StateT, G, P >;
 
     struct ExpandVisitor : public node_visitor_type
     {
-        explicit ExpandVisitor( Player &player ) noexcept : player( player ) {}
+        explicit ExpandVisitor( Player& player ) noexcept : player( player ) {}
 
         // thread-safe.
-        void visit( pre_node_type &node ) override
+        void visit( pre_node_type& node ) override
         {
             boost::intrusive::list< node_type > new_children;
-            for ( MoveT const &move : node.get_game() )
+            for ( MoveT const& move : node.get_game() )
             {
-                auto &child = *(
+                auto& child = *(
                     new ( player.allocator.allocate< pre_node_type >() )
                         pre_node_type( node.get_game().apply( move ), move ) );
 
@@ -209,15 +210,15 @@ class Player : public ::Player< MoveT >
             node.get_children().swap( new_children );
         }
 
-        Player &player;
+        Player& player;
     };
 
     struct InferenceVisitor : public node_visitor_type
     {
-        explicit InferenceVisitor( Player &player ) : player( player ) {}
+        explicit InferenceVisitor( Player& player ) : player( player ) {}
 
         // thread-safe.
-        void visit( pre_node_type &node ) noexcept override
+        void visit( pre_node_type& node ) noexcept override
         {
             // pushing a request to the inference service may block if the
             // request queue is full.
@@ -228,17 +229,17 @@ class Player : public ::Player< MoveT >
                   .state = player.serialize_state( node.get_game() ) } );
         }
 
-        Player &player;
+        Player& player;
     };
 
     struct SelectVisitor : public node_visitor_type
     {
-        explicit SelectVisitor( Player &player ) : player( player ) {}
+        explicit SelectVisitor( Player& player ) : player( player ) {}
 
         using base_node_type = Node< MoveT, StateT, payload_type >;
 
         // thread-safe.
-        void update_selection_stats( base_node_type &node )
+        void update_selection_stats( base_node_type& node )
         {
             if ( node.get_payload().policies_are_evaluated.load(
                      std::memory_order_relaxed ) )
@@ -249,13 +250,13 @@ class Player : public ::Player< MoveT >
 
         // require: child has children.
         // not thread-safe.
-        void select( base_node_type &node )
+        void select( base_node_type& node )
         {
             selected_node = &*std::ranges::max_element(
                 node.get_children(),
                 [this, parent_visits = node.get_payload().visits.load(
-                           std::memory_order_relaxed )]( auto const &a,
-                                                         auto const &b )
+                           std::memory_order_relaxed )]( auto const& a,
+                                                         auto const& b )
                 {
                     return player.ucb( a.get_payload(), parent_visits ) <
                            player.ucb( b.get_payload(), parent_visits );
@@ -263,7 +264,7 @@ class Player : public ::Player< MoveT >
         }
 
         // thread-safe because fix nodes are expanded already.
-        void visit( fix_node_type &node ) noexcept override
+        void visit( fix_node_type& node ) noexcept override
         {
             update_selection_stats( node );
             // fix nodes have children.
@@ -271,7 +272,7 @@ class Player : public ::Player< MoveT >
         }
 
         // thread-safe.
-        void visit( pre_node_type &node ) noexcept override
+        void visit( pre_node_type& node ) noexcept override
         {
             update_selection_stats( node );
 
@@ -284,14 +285,14 @@ class Player : public ::Player< MoveT >
             select( node );
         }
 
-        Player &player;
-        base_node_type *selected_node = nullptr;
+        Player& player;
+        base_node_type* selected_node = nullptr;
     };
 
     // thread-safe.
-    void simulation( node_type &node )
+    void simulation( node_type& node )
     {
-        auto &payload = node.get_payload();
+        auto& payload = node.get_payload();
 
         // Atomically update visit count. on first visit expand leaf node.
         const bool is_leaf_node = payload.visits.fetch_add( 1 ) == 0;
@@ -332,13 +333,13 @@ class Player : public ::Player< MoveT >
     {
         response_type response;
         while ( response_queue.pop( response ) )
-            backpropagation( *static_cast< node_type * >( response.node ),
+            backpropagation( *static_cast< node_type* >( response.node ),
                              response.policies, response.nn_value );
     }
 
     // require: single-threaded access.
     // return move to choose.
-    node_type &run_simulations()
+    node_type& run_simulations()
     {
         unfinished_simulations.store( game_play.simulations );
         // wake up waiting worker threads by initializing resource.
@@ -374,7 +375,7 @@ class Player : public ::Player< MoveT >
             // choose child with most visits
             return *std::ranges::max_element(
                 root.get().get_children(),
-                []( auto const &a, auto const &b )
+                []( auto const& a, auto const& b )
                 {
                     return a.get_payload().visits.load(
                                std::memory_order_relaxed ) <
@@ -385,7 +386,7 @@ class Player : public ::Player< MoveT >
 
     // upper confidence bound
     // thread-safe.
-    float ucb( payload_type const &payload, size_t parent_visits )
+    float ucb( payload_type const& payload, size_t parent_visits )
     {
         const size_t child_visits =
             payload.visits.load( std::memory_order_relaxed );
@@ -395,8 +396,8 @@ class Player : public ::Player< MoveT >
         float c =
             static_cast< float >( parent_visits ) + ucb_params.c_base + 1.0f;
         c /= ucb_params.c_base;
-        c = std::logf( c ) + ucb_params.c_init;
-        c *= std::sqrtf( static_cast< float >( parent_visits ) );
+        c = std::log( c ) + ucb_params.c_init;
+        c *= std::sqrt( static_cast< float >( parent_visits ) );
         c /= static_cast< float >( child_visits + 1 );
 
         float q = 0.0;
@@ -411,8 +412,8 @@ class Player : public ::Player< MoveT >
     // lockfree setting of policies and nn value and backpropagation.
     // require: mode is undecided.
     // thread-safe.
-    void backpropagation( node_type &node,
-                          std::array< float, P > const &policies,
+    void backpropagation( node_type& node,
+                          std::array< float, P > const& policies,
                           float nn_value )
     {
         // convert logits of legal moves to probabilities with softmax
@@ -421,17 +422,17 @@ class Player : public ::Player< MoveT >
         float policy_sum = 0.0f;
         std::array< float, P > policy_buffer;
         auto p_itr = policy_buffer.begin();
-        for ( node_type &child : node.get_children() )
+        for ( node_type& child : node.get_children() )
         {
             const float p =
-                std::expf( policies[move_to_policy_index( child.get_move() )] );
+                std::exp( policies[move_to_policy_index( child.get_move() )] );
             *p_itr++ = p;
             policy_sum += p;
         }
 
         // normalize priors to probabilities.
         p_itr = policy_buffer.begin();
-        for ( auto &child : node.get_children() )
+        for ( auto& child : node.get_children() )
             child.get_payload().nn_policy.store( *p_itr++ / policy_sum,
                                                  std::memory_order_relaxed );
 
@@ -444,7 +445,7 @@ class Player : public ::Player< MoveT >
 
         // backpropagate nn value. note the alternating sign of nn_value
         // the player's perspective is changing on the way up to the root node.
-        for ( node_type *next = &node; next; next = next->get_parent() )
+        for ( node_type* next = &node; next; next = next->get_parent() )
         {
             next->get_payload().nn_value_sum.fetch_add(
                 nn_value, std::memory_order_relaxed );
@@ -468,14 +469,14 @@ class Player : public ::Player< MoveT >
     }
 
     // require: root node is expanded
-    node_type &choose_opening_move()
+    node_type& choose_opening_move()
     {
-        auto &children = root.get().get_children();
+        auto& children = root.get().get_children();
 
         // sample from children in opening phase by visit distribution
         // so we are more versatile in the opening
         size_t total_visits = 0;
-        for ( auto const &child : children )
+        for ( auto const& child : children )
             total_visits +=
                 child.get_payload().visits.load( std::memory_order_relaxed );
 
@@ -501,7 +502,7 @@ class Player : public ::Player< MoveT >
         throw beat_it::Exception( "cannot choose move" );
     }
 
-    allocator_type &allocator;
+    allocator_type& allocator;
     const params::Ucb ucb_params;
     const params::GamePlay game_play;
     size_t move_count = 0;
@@ -512,7 +513,7 @@ class Player : public ::Player< MoveT >
     std::mutex simulations_mutex;
     std::condition_variable simulations_cond;
     std::stop_source stop_source;
-    inference_service_type &inference_service;
+    inference_service_type& inference_service;
     const float virtual_loss = 1.0; // must be positive
     boost::lockfree::queue< response_type > response_queue;
     std::reference_wrapper< node_type > root;
@@ -538,11 +539,11 @@ template < typename MoveT, typename StateT, size_t G, size_t P > class SelfPlay
     using expand_visitor_type = player_type::ExpandVisitor;
     using pre_node_type = player_type::pre_node_type;
 
-    SelfPlay( player_type &player, float dirichlet_alpha,
-              float dirichlet_epsilon, std::mt19937 &g,
-              std::vector< Position< G, P > > &positions,
-              Statistics &root_node_entropy_stat,
-              Statistics &informed_selection_stats ) noexcept
+    SelfPlay( player_type& player, float dirichlet_alpha,
+              float dirichlet_epsilon, std::mt19937& g,
+              std::vector< Position< G, P > >& positions,
+              Statistics& root_node_entropy_stat,
+              Statistics& informed_selection_stats ) noexcept
         : player( player ), dirichlet_epsilon( dirichlet_epsilon ), g( g ),
           gamma_dist( dirichlet_alpha, 1.0f ), positions( positions ),
           root_node_entropy_stat( root_node_entropy_stat ),
@@ -550,7 +551,7 @@ template < typename MoveT, typename StateT, size_t G, size_t P > class SelfPlay
     {
     }
 
-    Statistics const &get_root_node_entropy() const noexcept
+    Statistics const& get_root_node_entropy() const noexcept
     {
         return root_node_entropy_stat;
     }
@@ -563,7 +564,7 @@ template < typename MoveT, typename StateT, size_t G, size_t P > class SelfPlay
         // copied in the next generation root node due to optimization.
         Game< MoveT, StateT > game = [&root = player.get_root()]
         {
-            auto *pre_node = dynamic_cast< pre_node_type * >( &root );
+            auto* pre_node = dynamic_cast< pre_node_type* >( &root );
             if ( !pre_node )
                 throw beat_it::Exception( "invalid root node type" );
             return pre_node->get_game();
@@ -573,7 +574,7 @@ template < typename MoveT, typename StateT, size_t G, size_t P > class SelfPlay
               game_result == GameResult::Undecided;
               game_result = player.get_root().get_game_result() )
         {
-            auto &root = player.get_root();
+            auto& root = player.get_root();
             // expand root node on first visit, for dirichlet noise addition
             // the root node needs to be expanded.
             if ( root.get_children().empty() )
@@ -583,7 +584,7 @@ template < typename MoveT, typename StateT, size_t G, size_t P > class SelfPlay
             }
             add_dirichlet_noise();
 
-            auto &node = player.run_simulations();
+            auto& node = player.run_simulations();
             append_training_data( game );
             // update game.
             game = game.apply( node.get_move() );
@@ -613,10 +614,10 @@ template < typename MoveT, typename StateT, size_t G, size_t P > class SelfPlay
     // require: root node is expanded
     void add_dirichlet_noise()
     {
-        auto &root = player.get_root();
+        auto& root = player.get_root();
 
         // add noise for root node
-        for ( auto &child : root.get_children() )
+        for ( auto& child : root.get_children() )
         {
             float policy =
                 child.get_payload().nn_policy.load( std::memory_order_relaxed );
@@ -627,12 +628,12 @@ template < typename MoveT, typename StateT, size_t G, size_t P > class SelfPlay
         }
     }
 
-    void append_training_data( Game< MoveT, StateT > const &game )
+    void append_training_data( Game< MoveT, StateT > const& game )
     {
         positions.emplace_back();
-        auto &position = positions.back();
-        auto &root = player.get_root();
-        auto &payload = root.get_payload();
+        auto& position = positions.back();
+        auto& root = player.get_root();
+        auto& payload = root.get_payload();
 
         position.current_player = root.get_current_player_index();
 
@@ -649,7 +650,7 @@ template < typename MoveT, typename StateT, size_t G, size_t P > class SelfPlay
         // Then, iterate through the children once to populate the policies.
         // This is more efficient than iterating through all possible moves.
         if ( sum_visits > 0 )
-            for ( auto const &child : root.get_children() )
+            for ( auto const& child : root.get_children() )
             {
                 const size_t policy_index =
                     player.move_to_policy_index( child.get_move() );
@@ -661,13 +662,13 @@ template < typename MoveT, typename StateT, size_t G, size_t P > class SelfPlay
     }
 
   private:
-    Player< MoveT, StateT, G, P > &player;
+    Player< MoveT, StateT, G, P >& player;
     float dirichlet_epsilon;
-    std::mt19937 &g;
+    std::mt19937& g;
     std::gamma_distribution< float > gamma_dist;
-    std::vector< Position< G, P > > &positions;
-    Statistics &root_node_entropy_stat;
-    Statistics &informed_selection_stats;
+    std::vector< Position< G, P > >& positions;
+    Statistics& root_node_entropy_stat;
+    Statistics& informed_selection_stats;
 };
 
 } // namespace training
